COMMNT COZIR

+++ Data/Entities/OdersGoods.cs
@@ -2,11 +2,11 @@
 {
     public partial class OdersGoods
     {
-        public int Id { get; set; }
-        public int OrderId { get; set; }
+        public long Id { get; set; }
+        public long OrderId { get; set; }
         public Order Order { get; set; } = null!;
-        public int GoodId { get; set; }
+        public long GoodId { get; set; }
         public Good Good { get; set; } = null!;
-        public int CountGoods { get; set; }
+        public long CountGoods { get; set; }
     }
 }
