+++ Data/Entities/UsersGoods.cs
@@ -2,11 +2,11 @@
 {
     public partial class UsersGoods
     {
-        public int Id { get; set; }
-        public int UserId { get; set; }
+        public long Id { get; set; }
+        public long UserId { get; set; }
         public virtual User User { get; set; } = null!;
-        public int GoodId { get; set; }
+        public long GoodId { get; set; }
         public virtual Good Good { get; set; } = null!;
-        public int Remainder { get; set; } 
+        public long Remainder { get; set; } 
     }
 }
